refactor(epicenter): rename capabilities folder to extensions
Complete the rename from capabilities/ to extensions/ to better reflect
the type-safe workspace-level plugin system vs doc-level providers.

- Rename src/capabilities/ to src/extensions/
- Update package.json exports from /capabilities/* to /extensions/*
- Update all import paths across apps, examples, and scripts
- Update JSDoc comments and examples (capability → extension)
- Add revision-history to package.json exports
