+++ examples/content-hub/scripts/01-normalize-markdown.ts
@@ -14,7 +14,7 @@ import {
 	listMarkdownFiles,
 	readMarkdownFile,
 	writeMarkdownFile,
-} from '@epicenter/hq/capabilities/markdown';
+} from '@epicenter/hq/extensions/markdown';
 import { Err, Ok, type Result } from 'wellcrafted/result';
 
 const sourcePath = process.env.MARKDOWN_SOURCE_PATH;
