+++ packages/epicenter/scripts/email-minimal-simulation.ts
@@ -11,7 +11,7 @@
 
 import { existsSync, mkdirSync, rmSync, statSync } from 'node:fs';
 import { join } from 'node:path';
-import { persistence } from '../src/capabilities/persistence/desktop';
+import { persistence } from '../src/extensions/persistence/desktop';
 import {
 	createClient,
 	defineWorkspace,
