+++ packages/epicenter/scripts/email-storage-simulation.ts
@@ -15,7 +15,7 @@
 
 import { existsSync, mkdirSync, rmSync, statSync } from 'node:fs';
 import { join } from 'node:path';
-import { persistence } from '../src/capabilities/persistence/desktop';
+import { persistence } from '../src/extensions/persistence/desktop';
 import {
 	createClient,
 	defineWorkspace,
