+++ packages/epicenter/scripts/yjs-vs-sqlite-comparison.ts
@@ -12,7 +12,7 @@
 import { Database } from 'bun:sqlite';
 import { existsSync, mkdirSync, rmSync, statSync } from 'node:fs';
 import { join } from 'node:path';
-import { persistence } from '../src/capabilities/persistence/desktop';
+import { persistence } from '../src/extensions/persistence/desktop';
 import {
 	createClient,
 	defineWorkspace,
